+++ packages/image/__tests__/web/jest.config.js
@@ -1,5 +1,5 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
 
 module.exports = jestConfigurator("web", __dirname, {
-  coverageIgnoreGlobs: ["index.js", "modal-image"]
+  coverageIgnoreGlobs: ["index.js"]
 });