+++ packages/pull-quote/__tests__/shared-with-style.js
@@ -14,9 +14,6 @@ import shared from "./shared-with-style.base";
 jest.mock("@times-components/link", () => ({
   TextLink: "TextLink"
 }));
-jest.mock("@times-components/link", () => ({
-  TimesTextLink: "TimesTextLink"
-}));
 jest.mock("@times-components/icons", () => ({
   IconTwitter: "IconTwitter"
 }));