+++ packages/section/__tests__/shared.js
@@ -84,7 +84,7 @@ export default () => {
           onPuzzlePress={() => {}}
           publicationName="TIMES"
           recentlyOpenedPuzzleCount={1}
-          section={edition.sections[5]}
+          section={edition.sections[4]}
         />
       ).toJSON()
     ).toMatchSnapshot();
@@ -117,7 +117,7 @@ export default () => {
           onPuzzlePress={() => {}}
           publicationName="TIMES"
           recentlyOpenedPuzzleCount={1}
-          section={edition.sections[5]}
+          section={edition.sections[4]}
         />
       ).toJSON()
     ).toMatchSnapshot();
@@ -149,40 +149,6 @@ export default () => {
         <SectionContext.Provider value={{ recentlyOpenedPuzzleCount: 3 }}>
           <PuzzleBar onPress={() => {}} />
         </SectionContext.Provider>
-      ).toJSON()
-    ).toMatchSnapshot();
-  });
-
-  it("Times magazine section", () => {
-    const edition = new MockEdition().get();
-
-    expect(
-      TestRenderer.create(
-        <Section
-          analyticsStream={() => {}}
-          onArticlePress={() => {}}
-          onPuzzleBarPress={() => {}}
-          onPuzzlePress={() => {}}
-          publicationName="TIMES"
-          section={edition.sections[4]}
-        />
-      ).toJSON()
-    ).toMatchSnapshot();
-  });
-
-  it("Sunday Times magazine section", () => {
-    const edition = new MockEdition().get();
-
-    expect(
-      TestRenderer.create(
-        <Section
-          analyticsStream={() => {}}
-          onArticlePress={() => {}}
-          onPuzzleBarPress={() => {}}
-          onPuzzlePress={() => {}}
-          publicationName="SUNDAY TIMES"
-          section={edition.sections[4]}
-        />
       ).toJSON()
     ).toMatchSnapshot();
   });