+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.7.5](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.7.4...@times-components/related-articles@6.7.5) (2022-04-08)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
+## [6.7.4](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.7.3...@times-components/related-articles@6.7.4) (2022-04-08)
+
+**Note:** Version bump only for package @times-components/related-articles
+
+
+
+
+
 ## [6.7.3](https://github.com/newsuk/times-components/compare/@times-components/related-articles@6.7.2...@times-components/related-articles@6.7.3) (2022-04-05)
 
 **Note:** Version bump only for package @times-components/related-articles